pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zapret-hub"
version = "1.4.2"
description = "Windows desktop hub for managing external network utilities and configuration packs."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.7",
    "certifi>=2024.0.0",
    "cryptography>=46.0.0",
    "pydivert>=3.1.1",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.6",
    "nuitka>=2.7",
    "Pillow>=10.0",
]

[project.scripts]
zapret-hub = "zapret_hub.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
